Refactor browser and table into folders
